# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pricedug API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface Category {
  id: number;
  name: string;
  createdAt: string;
}

export interface CreateCategoryInput {
  name: string;
}

export interface Business {
  id: number;
  clerkUserId: string;
  name: string;
  description?: string | null;
  address?: string | null;
  phone?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  imageUrl?: string | null;
  isHidden: boolean;
  createdAt: string;
}

export interface CreateBusinessInput {
  name: string;
  description?: string | null;
  address?: string | null;
  phone?: string | null;
  categoryId?: number | null;
  imageUrl?: string | null;
}

export interface UpdateBusinessInput {
  name?: string;
  description?: string | null;
  address?: string | null;
  phone?: string | null;
  categoryId?: number | null;
  imageUrl?: string | null;
}

export interface VisibilityInput {
  isHidden: boolean;
}

export interface Product {
  id: number;
  businessId: number;
  name: string;
  description?: string | null;
  price?: string | null;
  imageUrl?: string | null;
  size?: string | null;
  materials?: string | null;
  createdAt: string;
}

export interface CreateProductInput {
  name: string;
  description?: string | null;
  price?: string | null;
  imageUrl?: string | null;
  size?: string | null;
  materials?: string | null;
}

export interface UpdateProductInput {
  name?: string;
  description?: string | null;
  price?: string | null;
  imageUrl?: string | null;
  size?: string | null;
  materials?: string | null;
}

export interface UploadUrlInput {
  filename: string;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadUrl: string;
  publicUrl: string;
}

export type GetBusinessesParams = {
categoryId?: number;
};