# lib/api-zod/src/generated/types/createBusinessInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pricedug API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBusinessInput {
  name: string;
  description?: string | null;
  address?: string | null;
  phone?: string | null;
  categoryId?: number | null;
  imageUrl?: string | null;
}