# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pricedug API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  businessId: number;
  name: string;
  description?: string | null;
  price?: string | null;
  imageUrl?: string | null;
  size?: string | null;
  materials?: string | null;
  createdAt: Date;
}