# lib/api-zod/src/generated/types/uploadUrlInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pricedug API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadUrlInput {
  filename: string;
  contentType: string;
}